Move all sources into a sources subpackage
Sources mixed in with utility and other helper-packages are hard to find
when we have so many.

Move all the packages specifically for sources under a new sources
package.

Command used for this rewrite in a temporary GOPATH (after a go mod
vendor):

for src in file env flag pflag json yaml toml static; do GO111MODULE=off gomvpkg -from github.com/vimeo/dials/$src -to github.com/vimeo/dials/sources/$src -vcs_mv_cmd "git mv {{.Src}} {{.Dst}}"; done
