+++ sources/json/json_test.go
@@ -1,33 +1,32 @@
-package yaml
+package json
 
 import (
 	"context"
 	"net"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/vimeo/dials"
-	"github.com/vimeo/dials/static"
+	"github.com/vimeo/dials/sources/static"
 )
 
-func TestYAML(t *testing.T) {
+func TestJSON(t *testing.T) {
 	type testConfig struct {
 		Val1 string
 		C    chan struct{}
 		Val2 int
 	}
-	yamlData := `---
-        val1: something
-        val2: 42
-`
+	jsonData := `{
+        "val1": "something",
+        "val2": 42
+    }`
 
 	myConfig := &testConfig{}
 	d, err := dials.Config(
 		context.Background(),
 		myConfig,
-		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
+		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
 	)
 	require.NoError(t, err)
 
@@ -37,7 +36,44 @@ func TestYAML(t *testing.T) {
 	assert.Equal(t, 42, c.Val2)
 }
 
-func TestShallowlyNestedYAML(t *testing.T) {
+func TestShallowlyNestedJSON(t *testing.T) {
+	type testConfig struct {
+		DatabaseName    string `dials:"database_name"`
+		DatabaseAddress string `dials:"database_address"`
+		DatabaseUser    struct {
+			Username  string `dials:"username"`
+			Password  string `dials:"password"`
+			IPAddress net.IP `dials:"ip_address"`
+		} `dials:"database_user"`
+	}
+
+	jsonData := `{
+        "database_name": "something",
+		"database_address": "127.0.0.1",
+		"database_user": {
+			"username": "test",
+			"password": "password",
+			"ip_address": "123.10.11.121"
+		}
+    }`
+
+	myConfig := &testConfig{}
+	d, err := dials.Config(
+		context.Background(),
+		myConfig,
+		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
+	)
+	require.NoError(t, err)
+
+	c := d.View()
+
+	assert.Equal(t, "something", c.DatabaseName)
+	assert.Equal(t, "test", c.DatabaseUser.Username)
+	assert.Equal(t, "password", c.DatabaseUser.Password)
+	assert.Equal(t, net.IPv4(123, 10, 11, 121), c.DatabaseUser.IPAddress)
+}
+
+func TestDeeplyNestedJSON(t *testing.T) {
 	type testConfig struct {
 		DatabaseName    string `dials:"database_name"`
 		DatabaseAddress string `dials:"database_address"`
@@ -45,14 +81,13 @@ func TestShallowlyNestedYAML(t *testing.T) {
 			Username   string `dials:"username"`
 			Password   string `dials:"password"`
 			OtherStuff struct {
-				Something string        `dials:"something"`
-				IPAddress net.IP        `dials:"ip_address"`
-				Timeout   time.Duration `dials:"timeout"`
+				Something string `dials:"something"`
+				IPAddress net.IP `dials:"ip_address"`
 			} `dials:"other_stuff"`
 		} `dials:"database_user"`
 	}
 
-	yamlData := `{
+	jsonData := `{
 	    "database_name": "something",
 		"database_address": "127.0.0.1",
 		"database_user": {
@@ -60,22 +95,19 @@ func TestShallowlyNestedYAML(t *testing.T) {
 			"password": "password",
 			"other_stuff": {
 				"something": "asdf",
-				"ip_address": "123.10.11.121",
-				"timeout": "10s",
+				"ip_address": "123.10.11.121"
 			}
 		}
 	}`
 
 	myConfig := &testConfig{}
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 	d, err := dials.Config(
-		ctx,
+		context.Background(),
 		myConfig,
-		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
+		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
 	)
-
 	require.NoError(t, err)
+
 	c := d.View()
 
 	assert.Equal(t, "something", c.DatabaseName)
@@ -83,10 +115,10 @@ func TestShallowlyNestedYAML(t *testing.T) {
 	assert.Equal(t, "password", c.DatabaseUser.Password)
 	assert.Equal(t, "asdf", c.DatabaseUser.OtherStuff.Something)
 	assert.Equal(t, net.IPv4(123, 10, 11, 121), c.DatabaseUser.OtherStuff.IPAddress)
-	assert.Equal(t, time.Duration(10*time.Second), c.DatabaseUser.OtherStuff.Timeout)
+
 }
 
-func TestMoreDeeplyNestedYAML(t *testing.T) {
+func TestMoreDeeplyNestedJSON(t *testing.T) {
 	type testConfig struct {
 		DatabaseName    string `dials:"database_name"`
 		DatabaseAddress string `dials:"database_address"`
@@ -95,15 +127,14 @@ func TestMoreDeeplyNestedYAML(t *testing.T) {
 			Password   string `dials:"password"`
 			OtherStuff struct {
 				Something struct {
-					AnotherField string        `dials:"another_field"`
-					IPAddress    net.IP        `dials:"ip_address"`
-					Timeout      time.Duration `dials:"timeout"`
+					AnotherField string `dials:"another_field"`
+					IPAddress    net.IP `dials:"ip_address"`
 				} `dials:"something"`
 			} `dials:"other_stuff"`
 		} `dials:"database_user"`
 	}
 
-	yamlData := `{
+	jsonData := `{
 	    "database_name": "something",
 		"database_address": "127.0.0.1",
 		"database_user": {
@@ -112,8 +143,7 @@ func TestMoreDeeplyNestedYAML(t *testing.T) {
 			"other_stuff": {
 				"something": {
 					"another_field": "asdf",
-					"ip_address": "123.10.11.121",
-					"timeout": "10s", 
+					"ip_address": "123.10.11.121"
 				}
 			}
 		}
@@ -123,7 +153,7 @@ func TestMoreDeeplyNestedYAML(t *testing.T) {
 	d, err := dials.Config(
 		context.Background(),
 		myConfig,
-		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
+		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
 	)
 	require.NoError(t, err)
 
@@ -134,25 +164,5 @@ func TestMoreDeeplyNestedYAML(t *testing.T) {
 	assert.Equal(t, "password", c.DatabaseUser.Password)
 	assert.Equal(t, "asdf", c.DatabaseUser.OtherStuff.Something.AnotherField)
 	assert.Equal(t, net.IPv4(123, 10, 11, 121), c.DatabaseUser.OtherStuff.Something.IPAddress)
-	assert.Equal(t, time.Duration(10*time.Second), c.DatabaseUser.OtherStuff.Something.Timeout)
-}
 
-func TestDecoderBadMarkup(t *testing.T) {
-	type testConfig struct {
-		Val1 string
-		C    chan struct{}
-		Val2 int
-	}
-	yamlData := `---
-        val1 something
-        val 2: 42
-`
-
-	myConfig := &testConfig{}
-	_, err := dials.Config(
-		context.Background(),
-		myConfig,
-		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
-	)
-	require.Error(t, err)
 }
