+++ sources/static/static.go
