+++ sources/yaml/yaml_test.go
@@ -1,32 +1,33 @@
-package json
+package yaml
 
 import (
 	"context"
 	"net"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/vimeo/dials"
-	"github.com/vimeo/dials/static"
+	"github.com/vimeo/dials/sources/static"
 )
 
-func TestJSON(t *testing.T) {
+func TestYAML(t *testing.T) {
 	type testConfig struct {
 		Val1 string
 		C    chan struct{}
 		Val2 int
 	}
-	jsonData := `{
-        "val1": "something",
-        "val2": 42
-    }`
+	yamlData := `---
+        val1: something
+        val2: 42
+`
 
 	myConfig := &testConfig{}
 	d, err := dials.Config(
 		context.Background(),
 		myConfig,
-		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
+		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
 	)
 	require.NoError(t, err)
 
@@ -36,44 +37,7 @@ func TestJSON(t *testing.T) {
 	assert.Equal(t, 42, c.Val2)
 }
 
-func TestShallowlyNestedJSON(t *testing.T) {
-	type testConfig struct {
-		DatabaseName    string `dials:"database_name"`
-		DatabaseAddress string `dials:"database_address"`
-		DatabaseUser    struct {
-			Username  string `dials:"username"`
-			Password  string `dials:"password"`
-			IPAddress net.IP `dials:"ip_address"`
-		} `dials:"database_user"`
-	}
-
-	jsonData := `{
-        "database_name": "something",
-		"database_address": "127.0.0.1",
-		"database_user": {
-			"username": "test",
-			"password": "password",
-			"ip_address": "123.10.11.121"
-		}
-    }`
-
-	myConfig := &testConfig{}
-	d, err := dials.Config(
-		context.Background(),
-		myConfig,
-		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
-	)
-	require.NoError(t, err)
-
-	c := d.View()
-
-	assert.Equal(t, "something", c.DatabaseName)
-	assert.Equal(t, "test", c.DatabaseUser.Username)
-	assert.Equal(t, "password", c.DatabaseUser.Password)
-	assert.Equal(t, net.IPv4(123, 10, 11, 121), c.DatabaseUser.IPAddress)
-}
-
-func TestDeeplyNestedJSON(t *testing.T) {
+func TestShallowlyNestedYAML(t *testing.T) {
 	type testConfig struct {
 		DatabaseName    string `dials:"database_name"`
 		DatabaseAddress string `dials:"database_address"`
@@ -81,13 +45,14 @@ func TestDeeplyNestedJSON(t *testing.T) {
 			Username   string `dials:"username"`
 			Password   string `dials:"password"`
 			OtherStuff struct {
-				Something string `dials:"something"`
-				IPAddress net.IP `dials:"ip_address"`
+				Something string        `dials:"something"`
+				IPAddress net.IP        `dials:"ip_address"`
+				Timeout   time.Duration `dials:"timeout"`
 			} `dials:"other_stuff"`
 		} `dials:"database_user"`
 	}
 
-	jsonData := `{
+	yamlData := `{
 	    "database_name": "something",
 		"database_address": "127.0.0.1",
 		"database_user": {
@@ -95,19 +60,22 @@ func TestDeeplyNestedJSON(t *testing.T) {
 			"password": "password",
 			"other_stuff": {
 				"something": "asdf",
-				"ip_address": "123.10.11.121"
+				"ip_address": "123.10.11.121",
+				"timeout": "10s",
 			}
 		}
 	}`
 
 	myConfig := &testConfig{}
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
 	d, err := dials.Config(
-		context.Background(),
+		ctx,
 		myConfig,
-		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
+		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
 	)
-	require.NoError(t, err)
 
+	require.NoError(t, err)
 	c := d.View()
 
 	assert.Equal(t, "something", c.DatabaseName)
@@ -115,10 +83,10 @@ func TestDeeplyNestedJSON(t *testing.T) {
 	assert.Equal(t, "password", c.DatabaseUser.Password)
 	assert.Equal(t, "asdf", c.DatabaseUser.OtherStuff.Something)
 	assert.Equal(t, net.IPv4(123, 10, 11, 121), c.DatabaseUser.OtherStuff.IPAddress)
-
+	assert.Equal(t, time.Duration(10*time.Second), c.DatabaseUser.OtherStuff.Timeout)
 }
 
-func TestMoreDeeplyNestedJSON(t *testing.T) {
+func TestMoreDeeplyNestedYAML(t *testing.T) {
 	type testConfig struct {
 		DatabaseName    string `dials:"database_name"`
 		DatabaseAddress string `dials:"database_address"`
@@ -127,14 +95,15 @@ func TestMoreDeeplyNestedJSON(t *testing.T) {
 			Password   string `dials:"password"`
 			OtherStuff struct {
 				Something struct {
-					AnotherField string `dials:"another_field"`
-					IPAddress    net.IP `dials:"ip_address"`
+					AnotherField string        `dials:"another_field"`
+					IPAddress    net.IP        `dials:"ip_address"`
+					Timeout      time.Duration `dials:"timeout"`
 				} `dials:"something"`
 			} `dials:"other_stuff"`
 		} `dials:"database_user"`
 	}
 
-	jsonData := `{
+	yamlData := `{
 	    "database_name": "something",
 		"database_address": "127.0.0.1",
 		"database_user": {
@@ -143,7 +112,8 @@ func TestMoreDeeplyNestedJSON(t *testing.T) {
 			"other_stuff": {
 				"something": {
 					"another_field": "asdf",
-					"ip_address": "123.10.11.121"
+					"ip_address": "123.10.11.121",
+					"timeout": "10s", 
 				}
 			}
 		}
@@ -153,7 +123,7 @@ func TestMoreDeeplyNestedJSON(t *testing.T) {
 	d, err := dials.Config(
 		context.Background(),
 		myConfig,
-		&static.StringSource{Data: jsonData, Decoder: &Decoder{}},
+		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
 	)
 	require.NoError(t, err)
 
@@ -164,5 +134,25 @@ func TestMoreDeeplyNestedJSON(t *testing.T) {
 	assert.Equal(t, "password", c.DatabaseUser.Password)
 	assert.Equal(t, "asdf", c.DatabaseUser.OtherStuff.Something.AnotherField)
 	assert.Equal(t, net.IPv4(123, 10, 11, 121), c.DatabaseUser.OtherStuff.Something.IPAddress)
+	assert.Equal(t, time.Duration(10*time.Second), c.DatabaseUser.OtherStuff.Something.Timeout)
+}
 
+func TestDecoderBadMarkup(t *testing.T) {
+	type testConfig struct {
+		Val1 string
+		C    chan struct{}
+		Val2 int
+	}
+	yamlData := `---
+        val1 something
+        val 2: 42
+`
+
+	myConfig := &testConfig{}
+	_, err := dials.Config(
+		context.Background(),
+		myConfig,
+		&static.StringSource{Data: yamlData, Decoder: &Decoder{}},
+	)
+	require.Error(t, err)
 }
