+++ sourcewrap/transforming_source_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/vimeo/dials"
-	"github.com/vimeo/dials/static"
+	"github.com/vimeo/dials/sources/static"
 	"github.com/vimeo/dials/transform"
 )
